pyproject.toml
[project]
name = "workout-tracker"
version = "0.1.0"
description = "A simple workout tracker app"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "pandas-stubs~=2.3.3",
    "redis>=5.0.0",
    "pyjwt>=2.8.0",
    "slowapi>=0.1.9",
    "arq>=0.25.0",
    "logfire>=0.41.0",
    "mcp[cli]>=1.1.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "pytest-anyio>=0.0.0",
    "anyio>=4.0.0",
    "schemathesis>=4.9.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "pytest-anyio>=0.0.0",
    "anyio>=4.0.0",
    "schemathesis>=4.9.5",
]

